Move and rename 'beatmapdownloader' package to top level as 'beatmaplisting'
